defined getThresholdLevel for BroadcastLogger and moved persistent logger and plugin to devmenu package
